refactor(host): split runtime stop route coverage
